Move mount and update tree functions to internal
